hisq_stencil_test now runs via gtest, creating a ctest is outstanding
