+++ tests/hisq_stencil_test.cpp
@@ -1,435 +1,75 @@
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-
-#include <quda.h>
-#include <gauge_field.h>
-#include <host_utils.h>
-#include <command_line_params.h>
-#include <misc.h>
-#include <unitarization_links.h>
-#include <ks_improved_force.h>
-
-#include <tune_quda.h>
-
-#define TDIFF(a, b) (b.tv_sec - a.tv_sec + 0.000001 * (b.tv_usec - a.tv_usec))
+#include "hisq_stencil_test_utils.h"
 
 using namespace quda;
 
-// Number of naiks. If eps_naik is 0.0, we only need
-// to construct one naik.
-static QudaGaugeFieldOrder gauge_order = QUDA_MILC_GAUGE_ORDER;
+class HisqStencilTest : public ::testing::Test
+{
+protected:
+  HisqStencilTestWrapper hisq_stencil_test_wrapper;
 
-// The file "generic_ks/fermion_links_hisq_load_milc.c"
-// within MILC is the ultimate reference for what's going on here.
-
-// Unitarization coefficients
-static double unitarize_eps = 1e-6;
-static bool reunit_allow_svd = true;
-static bool reunit_svd_only = false;
-static double svd_rel_error = 1e-4;
-static double svd_abs_error = 1e-4;
-static double max_allowed_error = 1e-11;
-
-struct HisqStencilTestWrapper {
-
-  static inline QudaGaugeParam gauge_param;
-
-  // staple coefficients for different portions of the HISQ stencil build
-  static inline std::array<std::array<double, 6>, 3> act_paths;
-
-  // initial links in MILC order
-  static inline void* milc_sitelink = nullptr;
-
-  // storage for CPU reference fat and long links w/zero Naik
-  static inline void *fat_reflink[4] = {nullptr, nullptr, nullptr, nullptr};
-  static inline void *long_reflink[4] = {nullptr, nullptr, nullptr, nullptr};
-
-  // storage for CPU reference fat and long links w/non-zero Naik
-  static inline void *fat_reflink_eps[4] = {nullptr, nullptr, nullptr, nullptr};
-  static inline void *long_reflink_eps[4] = {nullptr, nullptr, nullptr, nullptr};
-
-  // Paths for step 1:
-  static inline void *vlink = nullptr;
-  static inline void *wlink = nullptr;
-
-  // Paths for step 2:
-  static inline void *fatlink = nullptr;
-  static inline void *longlink = nullptr;
-
-  // Place to accumulate Naiks
-  static inline void *fatlink_eps = nullptr;
-  static inline void *longlink_eps = nullptr;
-
-  static inline void *qdp_fatlink[4] = {nullptr, nullptr, nullptr, nullptr};
-  static inline void *qdp_longlink[4] = {nullptr, nullptr, nullptr, nullptr};
-  static inline void *qdp_fatlink_eps[4] = {nullptr, nullptr, nullptr, nullptr};
-  static inline void *qdp_longlink_eps[4] = {nullptr, nullptr, nullptr, nullptr};
-
-  void init_test() {
-    cpu_prec = prec;
-    host_gauge_data_type_size = cpu_prec;
-
-    gauge_param = newQudaGaugeParam();
-    setStaggeredGaugeParam(gauge_param);
-
-    static bool first_time = true;
-    if (first_time) {
-      init_host();
-      first_time = false;
-    }
+  void display_test_info() {
+    printfQuda("running the following test:\n");
+    printfQuda("link_precision           link_reconstruct           space_dimension        T_dimension       Ordering\n");
+    printfQuda("%s                       %s                         %d/%d/%d/                  %d             %s \n",
+              get_prec_str(prec), get_recon_str(link_recon), xdim, ydim, zdim, tdim, get_gauge_order_str(gauge_order));
+    printfQuda("Grid partition info:     X  Y  Z  T\n");
+    printfQuda("                         %d  %d  %d  %d\n", dimPartitioned(0), dimPartitioned(1), dimPartitioned(2),
+              dimPartitioned(3));
+    printfQuda("Number of Naiks: %d\n", n_naiks);
   }
 
-  void init_host() {
-    setDims(gauge_param.X);
-    dw_setDims(gauge_param.X, 1);
-
-    ///////////////////////////////////////////////////////////////
-    // Set up the coefficients for each part of the HISQ stencil //
-    ///////////////////////////////////////////////////////////////
-
-    // Reference: "generic_ks/imp_actions/hisq/hisq_action.h",
-    // in QHMC: https://github.com/jcosborn/qhmc/blob/master/lib/qopqdp/hisq.c
-
-    double u1 = 1.0 / tadpole_factor;
-    double u2 = u1 * u1;
-    double u4 = u2 * u2;
-    double u6 = u4 * u2;
-
-    // First path: create V, W links
-    act_paths[0] = {
-      (1.0 / 8.0),                             /* one link */
-      u2 * (0.0),                              /* Naik */
-      u2 * (-1.0 / 8.0) * 0.5,                 /* simple staple */
-      u4 * (1.0 / 8.0) * 0.25 * 0.5,           /* displace link in two directions */
-      u6 * (-1.0 / 8.0) * 0.125 * (1.0 / 6.0), /* displace link in three directions */
-      u4 * (0.0)                               /* Lepage term */
-    };
-
-    // Second path: create X, long links
-    act_paths[1] = {
-      ((1.0 / 8.0) + (2.0 * 6.0 / 16.0) + (1.0 / 8.0)), /* one link */
-                                                        /* One link is 1/8 as in fat7 + 2*3/8 for Lepage + 1/8 for Naik */
-      (-1.0 / 24.0),                                    /* Naik */
-      (-1.0 / 8.0) * 0.5,                               /* simple staple */
-      (1.0 / 8.0) * 0.25 * 0.5,                         /* displace link in two directions */
-      (-1.0 / 8.0) * 0.125 * (1.0 / 6.0),               /* displace link in three directions */
-      (-2.0 / 16.0)                                     /* Lepage term, correct O(a^2) 2x ASQTAD */
-    };
-
-    // Paths for epsilon corrections. Not used if n_naiks = 1.
-    act_paths[2] = {
-      (1.0 / 8.0),   /* one link b/c of Naik */
-      (-1.0 / 24.0), /* Naik */
-      0.0,           /* simple staple */
-      0.0,           /* displace link in two directions */
-      0.0,           /* displace link in three directions */
-      0.0            /* Lepage term */
-    };
-
-    ////////////////////////////////////
-    // Set unitarization coefficients //
-    ////////////////////////////////////
-
-    setUnitarizeLinksConstants(unitarize_eps, max_allowed_error, reunit_allow_svd, reunit_svd_only, svd_rel_error,
-                              svd_abs_error);
-
-    /////////////////
-    // Input links //
-    /////////////////
-
-    void *qdp_sitelink[4] = {nullptr, nullptr, nullptr, nullptr};
-    for (int i = 0; i < 4; i++) qdp_sitelink[i] = pinned_malloc(V * gauge_site_size * host_gauge_data_type_size);
-
-    milc_sitelink = (void *)safe_malloc(4 * V * gauge_site_size * host_gauge_data_type_size);
-
-    // Note: this could be replaced with loading a gauge field
-    createSiteLinkCPU(qdp_sitelink, gauge_param.cpu_prec, 0); // 0 -> no phases
-    reorderQDPtoMILC(milc_sitelink, qdp_sitelink, V, gauge_site_size, gauge_param.cpu_prec, gauge_param.cpu_prec);
-
-    ///////////////////////
-    // Perform CPU Build //
-    ///////////////////////
-
-    for (int i = 0; i < 4; i++) {
-      // fat and long links for fermions with zero epsilon
-      fat_reflink[i] = safe_malloc(V * gauge_site_size * host_gauge_data_type_size);
-      long_reflink[i] = safe_malloc(V * gauge_site_size * host_gauge_data_type_size);
-
-      // fat and long links for fermions with non-zero epsilon
-      if (n_naiks > 1) {
-        fat_reflink_eps[i] = safe_malloc(V * gauge_site_size * host_gauge_data_type_size);
-        long_reflink_eps[i] = safe_malloc(V * gauge_site_size * host_gauge_data_type_size);
-      }
-    }
-
-    computeHISQLinksCPU(fat_reflink, long_reflink, fat_reflink_eps, long_reflink_eps, qdp_sitelink, &gauge_param,
-                        act_paths, eps_naik);
-
-    ///////////////////////////////////////////////////////
-    // Allocate host storage for fields built on the GPU //
-    ///////////////////////////////////////////////////////
-
-    // Paths for step 1:
-    vlink = pinned_malloc(4 * V * gauge_site_size * host_gauge_data_type_size); // V links
-    wlink = pinned_malloc(4 * V * gauge_site_size * host_gauge_data_type_size); // W links
-
-    // Paths for step 2:
-    fatlink = pinned_malloc(4 * V * gauge_site_size * host_gauge_data_type_size);  // final fat ("X") links
-    longlink = pinned_malloc(4 * V * gauge_site_size * host_gauge_data_type_size); // final long links
-
-    // Place to accumulate Naiks
-    if (n_naiks > 1) {
-      fatlink_eps = pinned_malloc(4 * V * gauge_site_size * host_gauge_data_type_size);  // epsilon fat links
-      longlink_eps = pinned_malloc(4 * V * gauge_site_size * host_gauge_data_type_size); // epsilon long naiks
-    }
-
-    // QDP order fields
-    for (int i = 0; i < 4; i++) {
-      qdp_fatlink[i] = safe_malloc(V * gauge_site_size * host_gauge_data_type_size);
-      qdp_longlink[i] = safe_malloc(V * gauge_site_size * host_gauge_data_type_size);
-      if (n_naiks > 1) {
-        qdp_fatlink_eps[i] = safe_malloc(V * gauge_site_size * host_gauge_data_type_size);
-        qdp_longlink_eps[i] = safe_malloc(V * gauge_site_size * host_gauge_data_type_size);
-      }
-    }
-
-    /////////////////////////////////////////////////////////
-    // Free allocations that are only needed for CPU setup //
-    /////////////////////////////////////////////////////////
-
-    for (int i = 0; i < 4; i++)
-      host_free(qdp_sitelink[i]);
+public:
+  virtual void SetUp() {
+    hisq_stencil_test_wrapper.init_test();
+    display_test_info();
   }
 
-  static void destroy() {
-    if (milc_sitelink) host_free(milc_sitelink);
-
-    for (int i = 0; i < 4; i++) {
-      host_free(fat_reflink[i]);
-      host_free(long_reflink[i]);
-      if (n_naiks > 1) {
-        host_free(fat_reflink_eps[i]);
-        host_free(long_reflink_eps[i]);
-      }
-    }
-
-    // Clean up GPU compute links
-    host_free(vlink);
-    host_free(wlink);
-    host_free(fatlink);
-    host_free(longlink);
-
-    if (n_naiks > 1) {
-      host_free(fatlink_eps);
-      host_free(longlink_eps);
-    }
-
-    for (int i = 0; i < 4; i++) {
-      host_free(qdp_fatlink[i]);
-      host_free(qdp_longlink[i]);
-      if (n_naiks > 1) {
-        host_free(qdp_fatlink_eps[i]);
-        host_free(qdp_longlink_eps[i]);
-      }
-    }
-
-#ifdef MULTI_GPU
-    exchange_llfat_cleanup();
-#endif
+  virtual void TearDown() {
+    hisq_stencil_test_wrapper.end();
   }
 
-  /*--------------------------------------------------------------------*/
-  // Some notation:
-  // U -- original link, SU(3), copied to "field" from "site"
-  // V -- after 1st level of smearing, non-SU(3)
-  // W -- unitarized, SU(3)
-  // X -- after 2nd level of smearing, non-SU(3)
-  /*--------------------------------------------------------------------*/
-
-  double llfatCUDA(int niter) {
-    host_timer_t host_timer;
-
-    comm_barrier();
-    host_timer.start();
-
-    for (int i = 0; i < niter; i++) {
-      // If we create cudaGaugeField objs, we can do this 100% on the GPU, no copying!
-
-      // Create V links (fat7 links) and W links (unitarized V links), 1st path table set
-      computeKSLinkQuda(vlink, nullptr, wlink, milc_sitelink, act_paths[0].data(), &gauge_param);
-
-      if (n_naiks > 1) {
-        // Create Naiks, 3rd path table set
-        computeKSLinkQuda(fatlink, longlink, nullptr, wlink, act_paths[2].data(), &gauge_param);
-
-        // Rescale+copy Naiks into Naik field
-        cpu_axy(prec, eps_naik, fatlink, fatlink_eps, V * 4 * gauge_site_size);
-        cpu_axy(prec, eps_naik, longlink, longlink_eps, V * 4 * gauge_site_size);
-      } else {
-        memset(fatlink, 0, V * 4 * gauge_site_size * host_gauge_data_type_size);
-        memset(longlink, 0, V * 4 * gauge_site_size * host_gauge_data_type_size);
-      }
-
-      // Create X and long links, 2nd path table set
-      computeKSLinkQuda(fatlink, longlink, nullptr, wlink, act_paths[1].data(), &gauge_param);
-
-      if (n_naiks > 1) {
-        // Add into Naik field
-        cpu_xpy(prec, fatlink, fatlink_eps, V * 4 * gauge_site_size);
-        cpu_xpy(prec, longlink, longlink_eps, V * 4 * gauge_site_size);
-      }
-    }
-
-    host_timer.stop();
-
-    return host_timer.last();
+  static void SetUpTestCase() {
+    initQuda(device_ordinal);
   }
 
-  void run_test(int niter, bool print_metrics = false) {
-    //////////////////////
-    // Perform GPU test //
-    //////////////////////
-
-    printfQuda("Tuning...\n");
-    llfatCUDA(1);
-
-    auto flops0 = quda::Tunable::flops_global();
-    auto bytes0 = quda::Tunable::bytes_global();
-
-    printfQuda("Running %d iterations of computation\n", niter);
-    double secs = llfatCUDA(niter);
-
-    unsigned long long flops = (quda::Tunable::flops_global() - flops0);
-    unsigned long long bytes = (quda::Tunable::bytes_global() - bytes0);
-
-    if (print_metrics) {
-      // FIXME: does not include unitarization, extra naiks
-      int volume = gauge_param.X[0] * gauge_param.X[1] * gauge_param.X[2] * gauge_param.X[3];
-      //long long flops = 61632 * (long long)niter; // Constructing V field
-      // Constructing W field?
-      // Constructing separate Naiks
-      //flops += 61632 * (long long)niter;     // Constructing X field
-      //flops += (252 * 4) * (long long)niter; // long-link contribution
-
-      printfQuda("%fus per HISQ link build\n", 1e6 * secs / niter);
-
-      printfQuda("%llu flops per HISQ link build, %llu flops per site %llu bytes per site\n", flops / niter,
-                    (flops / niter) / volume, (bytes / niter) / volume);
-
-      double gflops = 1.0e-9 * flops / secs;
-        printfQuda("GFLOPS = %f\n", gflops);
-
-      double gbytes = 1.0e-9 * bytes / secs;
-        printfQuda("GBYTES = %f\n", gbytes);
-
-      // Old metric
-      //double perf = flops / (secs * 1024 * 1024 * 1024);
-      //printfQuda("link computation time =%.2f ms, flops= %.2f Gflops\n", (secs * 1000) / niter, perf);
-    }
-  }
-
-  void verify()
+  // Per-test-case tear-down.
+  // Called after the last test in this test case.
+  // Can be omitted if not needed.
+  static void TearDownTestCase()
   {
-    ////////////////////////////////////////////////////////////////////
-    // Layout change for fatlink, fatlink_eps, longlink, longlink_eps //
-    ////////////////////////////////////////////////////////////////////
-
-    reorderMILCtoQDP(qdp_fatlink, fatlink, V, gauge_site_size, gauge_param.cpu_prec, gauge_param.cpu_prec);
-    reorderMILCtoQDP(qdp_longlink, longlink, V, gauge_site_size, gauge_param.cpu_prec, gauge_param.cpu_prec);
-
-    if (n_naiks > 1) {
-      reorderMILCtoQDP(qdp_fatlink_eps, fatlink_eps, V, gauge_site_size, gauge_param.cpu_prec, gauge_param.cpu_prec);
-      reorderMILCtoQDP(qdp_longlink_eps, longlink_eps, V, gauge_site_size, gauge_param.cpu_prec, gauge_param.cpu_prec);
-    }
-
-    //////////////////////////////
-    // Perform the verification //
-    //////////////////////////////
-
-    printfQuda("Checking fat links...\n");
-    int res = 1;
-    for (int dir = 0; dir < 4; dir++) {
-      res &= compare_floats(fat_reflink[dir], qdp_fatlink[dir], V * gauge_site_size, 1e-3, gauge_param.cpu_prec);
-    }
-
-    strong_check_link(qdp_fatlink, "GPU results: ", fat_reflink, "CPU reference results:", V, gauge_param.cpu_prec);
-
-    printfQuda("Fat-link test %s\n\n", (1 == res) ? "PASSED" : "FAILED");
-
-    printfQuda("Checking long links...\n");
-    res = 1;
-    for (int dir = 0; dir < 4; ++dir) {
-      res &= compare_floats(long_reflink[dir], qdp_longlink[dir], V * gauge_site_size, 1e-3, gauge_param.cpu_prec);
-    }
-
-    strong_check_link(qdp_longlink, "GPU results: ", long_reflink, "CPU reference results:", V, gauge_param.cpu_prec);
-
-    printfQuda("Long-link test %s\n\n", (1 == res) ? "PASSED" : "FAILED");
-
-    if (n_naiks > 1) {
-
-      printfQuda("Checking fat eps_naik links...\n");
-      res = 1;
-      for (int dir = 0; dir < 4; dir++) {
-        res &= compare_floats(fat_reflink_eps[dir], qdp_fatlink_eps[dir], V * gauge_site_size, 1e-3,
-                              gauge_param.cpu_prec);
-      }
-
-      strong_check_link(qdp_fatlink_eps, "GPU results: ", fat_reflink_eps, "CPU reference results:", V,
-                        gauge_param.cpu_prec);
-
-      printfQuda("Fat-link eps_naik test %s\n\n", (1 == res) ? "PASSED" : "FAILED");
-
-      printfQuda("Checking long eps_naik links...\n");
-      res = 1;
-      for (int dir = 0; dir < 4; ++dir) {
-        res &= compare_floats(long_reflink_eps[dir], qdp_longlink_eps[dir], V * gauge_site_size, 1e-3,
-                              gauge_param.cpu_prec);
-      }
-
-      strong_check_link(qdp_longlink_eps, "GPU results: ", long_reflink_eps, "CPU reference results:", V,
-                        gauge_param.cpu_prec);
-
-      printfQuda("Long-link eps_naik test %s\n\n", (1 == res) ? "PASSED" : "FAILED");
-    }
+    HisqStencilTestWrapper::destroy();
+    endQuda();
   }
 };
 
-static void hisq_test()
+TEST_F(HisqStencilTest, benchmark)
 {
-  initQuda(device_ordinal);
-
-  HisqStencilTestWrapper test_wrapper;
-
-  test_wrapper.init_test();
-
-  test_wrapper.run_test(niter, true);
-
-  test_wrapper.verify();
-
-  test_wrapper.destroy();
-
-  endQuda();
+  hisq_stencil_test_wrapper.run_test(niter, /**show_metrics =*/true);
 }
 
-static void display_test_info()
+TEST_F(HisqStencilTest, verify)
 {
-  printfQuda("running the following test:\n");
+  if (!verify_results) GTEST_SKIP();
 
-  printfQuda("link_precision           link_reconstruct           space_dimension        T_dimension       Ordering\n");
-  printfQuda("%s                       %s                         %d/%d/%d/                  %d             %s \n",
-             get_prec_str(prec), get_recon_str(link_recon), xdim, ydim, zdim, tdim, get_gauge_order_str(gauge_order));
+  hisq_stencil_test_wrapper.run_test(2);
 
-  printfQuda("Grid partition info:     X  Y  Z  T\n");
-  printfQuda("                         %d  %d  %d  %d\n", dimPartitioned(0), dimPartitioned(1), dimPartitioned(2),
-             dimPartitioned(3));
+  std::array<double, 2> res = hisq_stencil_test_wrapper.verify();
 
-  printfQuda("Number of Naiks: %d\n", n_naiks);
+  // extra factor of 10 b/c the norm isn't normalized
+  double max_dev = 10. * getTolerance(prec);
+
+  // fat link
+  EXPECT_LE(res[0], max_dev);
+
+  // long link
+  EXPECT_LE(res[1], max_dev);
 }
 
 int main(int argc, char **argv)
 {
+  // initalize google test
+  ::testing::InitGoogleTest(&argc, argv);
+
   // for speed
   xdim = ydim = zdim = tdim = 8;
 
@@ -437,11 +77,8 @@ int main(int argc, char **argv)
   link_recon = QUDA_RECONSTRUCT_NO;
   cpu_prec = prec = QUDA_DOUBLE_PRECISION;
 
+  // Parse command line options
   auto app = make_app();
-  // app->get_formatter()->column_width(40);
-  // add_eigen_option_group(app);
-  // add_deflation_option_group(app);
-  // add_multigrid_option_group(app);
   try {
     app->parse(argc, argv);
   } catch (const CLI::ParseError &e) {
@@ -456,9 +93,15 @@ int main(int argc, char **argv)
   if (eps_naik != 0.0) { n_naiks = 2; }
 
   setVerbosity(verbosity);
-
   initComms(argc, argv, gridsize_from_cmdline);
-  display_test_info();
-  hisq_test();
+
+  // Ensure gtest prints only from rank 0
+  ::testing::TestEventListeners &listeners = ::testing::UnitTest::GetInstance()->listeners();
+  if (comm_rank() != 0) { delete listeners.Release(listeners.default_result_printer()); }
+
+  int test_rc = RUN_ALL_TESTS();
+
   finalizeComms();
+
+  return test_rc;
 }
